rename: 로그인 type 정의 파일 이름 auth로 재설정 (#37)
