+++ src/components/auth/login/LoginForm.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { LoginFormPropsType } from '@/types/login';
+import { LoginFormPropsType } from '@/types/auth';
 import InputField from './InputField';
 import { BottomButton } from '../BottomButton';
 import RememberBox from './RememberBox';
